Move UI stuff to __main__
