+++ mailmerge/api.py
@@ -175,7 +175,7 @@ class MessageTemplate:
             # Check that the attachment exists
             if not os.path.exists(normalized_path):
                 print("Error: can't find attachment " + normalized_path)
-                sys.exit(1)
+                sys.exit(1)  # FIXME
 
             filename = os.path.basename(normalized_path)
             with open(normalized_path, "rb") as attachment:
@@ -235,80 +235,15 @@ class SendmailClient:
         smtp.close()
 
 
-def create_sample_input_files(template_filename,
-                              database_filename,
-                              config_filename):
-    """Create sample template email and database."""
-    print("Creating sample template email {}".format(template_filename))
-    if os.path.exists(template_filename):
-        print("Error: file exists: " + template_filename)
-        sys.exit(1)
-    with io.open(template_filename, "w") as template_file:
-        template_file.write(
-            u"TO: {{email}}\n"
-            u"SUBJECT: Testing mailmerge\n"
-            u"FROM: My Self <myself@mydomain.com>\n"
-            u"\n"
-            u"Hi, {{name}},\n"
-            u"\n"
-            u"Your number is {{number}}.\n"
-        )
-    print("Creating sample database {}".format(database_filename))
-    if os.path.exists(database_filename):
-        print("Error: file exists: " + database_filename)
-        sys.exit(1)
-    with io.open(database_filename, "w") as database_file:
-        database_file.write(
-            u'email,name,number\n'
-            u'myself@mydomain.com,"Myself",17\n'
-            u'bob@bobdomain.com,"Bob",42\n'
-        )
-    print("Creating sample config file {}".format(config_filename))
-    if os.path.exists(config_filename):
-        print("Error: file exists: " + config_filename)
-        sys.exit(1)
-    with io.open(config_filename, "w") as config_file:
-        config_file.write(
-            u"# Example: GMail\n"
-            u"[smtp_server]\n"
-            u"host = smtp.gmail.com\n"
-            u"port = 465\n"
-            u"security = SSL/TLS\n"
-            u"username = YOUR_USERNAME_HERE\n"
-            u"#\n"
-            u"# Example: Wide open\n"
-            u"# [smtp_server]\n"
-            u"# host = open-smtp.example.com\n"
-            u"# port = 25\n"
-            u"# security = Never\n"
-            u"# username = None\n"
-            u"#\n"
-            u"# Example: University of Michigan\n"
-            u"# [smtp_server]\n"
-            u"# host = smtp.mail.umich.edu\n"
-            u"# port = 465\n"
-            u"# security = SSL/TLS\n"
-            u"# username = YOUR_USERNAME_HERE\n"
-            u"#\n"
-            u"# Example: University of Michigan EECS Dept., with STARTTLS security\n"  # noqa: E501
-            u"# [smtp_server]\n"
-            u"# host = newman.eecs.umich.edu\n"
-            u"# port = 25\n"
-            u"# security = STARTTLS\n"
-            u"# username = YOUR_USERNAME_HERE\n"
-            u"#\n"
-            u"# Example: University of Michigan EECS Dept., with no encryption\n"  # noqa: E501
-            u"# [smtp_server]\n"
-            u"# host = newman.eecs.umich.edu\n"
-            u"# port = 25\n"
-            u"# security = Never\n"
-            u"# username = YOUR_USERNAME_HERE\n"
-        )
-    print("Edit these files, and then run mailmerge again")
+def read_database(database_filename):
+    """Read database CSV file, providing one line at a time."""
+    with io.open(database_filename, "r") as database_file:
+        reader = csv.DictReader(database_file)
+        for row in reader:
+            yield row
 
 
-def main(sample=False,
-         dry_run=True,
+def main(dry_run=True,
          limit=1,
          no_limit=False,
          database_filename=DATABASE_FILENAME_DEFAULT,
@@ -322,75 +257,25 @@ def main(sample=False,
 
     Render an email template for each line in a CSV database.
     """
-    # Create a sample email template and database if there isn't one already
-    if sample:
-        create_sample_input_files(
-            template_filename,
-            database_filename,
-            config_filename,
-        )
-        sys.exit(0)
-    if not os.path.exists(template_filename):
-        print("Error: can't find template email " + template_filename)
-        print("Create a sample (--sample) or specify a file (--template)")
-        sys.exit(1)
-    if not os.path.exists(database_filename):
-        print("Error: can't find database_filename " + database_filename)
-        print("Create a sample (--sample) or specify a file (--database)")
-        sys.exit(1)
+    # Read template
+    message_template = MessageTemplate(template_filename)
 
-    try:
-        # Read template
-        message_template = MessageTemplate(template_filename)
+    # Read CSV file database
+    database = read_database(database_filename)
 
-        # Read CSV file database
-        database = []
-        with io.open(database_filename, "r") as database_file:
-            reader = csv.DictReader(database_file)
-            for row in reader:
-                database.append(row)
+    # Read SMTP client configuration
+    sendmail_client = SendmailClient(config_filename)
 
-        # Read SMTP client configuration
-        sendmail_client = SendmailClient(config_filename)
+    # Each row corresponds to one email message
+    for i, row in enumerate(database):
+        if not no_limit and i >= limit:
+            break
 
-        # Each row corresponds to one email message
-        for i, row in enumerate(database):
-            if not no_limit and i >= limit:
-                break
+        sender, recipients, message = message_template.render(row)
+        print(">>> message {}".format(i))  # FIXME
+        print(message.as_string())
 
-            sender, recipients, message = message_template.render(row)
-            print(">>> message {}".format(i))
-            print(message.as_string())
-
-            # Send message
-            if dry_run:
-                print(">>> sent message {} DRY RUN".format(i))
-            else:
-                sendmail_client.sendmail(sender, recipients, message)
-                print(">>> sent message {}".format(i))
-
-        # Hints for user
-        if not message_template.attachments:
-            print(">>> No attachments were sent with the emails.")
-        if not no_limit:
-            print(">>> Limit was {} messages.  ".format(limit) +
-                  "To remove the limit, use the --no-limit option.")
-        if dry_run:
-            print((">>> This was a dry run.  "
-                   "To send messages, use the --no-dry-run option."))
-
-    except jinja2.exceptions.TemplateError as err:
-        print(">>> Error in Jinja2 template: {}".format(err))
-        sys.exit(1)
-    except csv.Error as err:
-        print(">>> Error reading CSV file: {}".format(err))
-        sys.exit(1)
-    except smtplib.SMTPAuthenticationError as err:
-        print(">>> Authentication error: {}".format(err))
-        sys.exit(1)
-    except configparser.Error as err:
-        print(">>> Error reading config file {}: {}".format(
-            config_filename, err))
-        sys.exit(1)
-    except smtplib.SMTPException as err:
-        print(">>> Error sending message", err, sep=' ', file=sys.stderr)
+        # Send message
+        if not dry_run:
+            sendmail_client.sendmail(sender, recipients, message)
+            print(">>> sent message {}".format(i))  # FIXME
